requirements_prod.txt
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.12.2
bs4==0.0.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.0
curl_cffi==0.13.0
cycler==0.12.1
Django==4.2.20
django-cors-headers==4.3.1
django-environ==0.4.5
django-filter==24.1
django-stubs==5.1.3
django-stubs-ext==5.1.3
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djangorestframework-stubs==3.15.3
djoser==2.0.3
exceptiongroup==1.3.1
fonttools==4.60.2
frozendict==2.4.7
git-filter-repo==2.47.0
gunicorn==23.0.0
h11==0.16.0
idna==3.11
importlib_resources==6.5.2
kiwisolver==1.4.7
lxml==4.9.3
matplotlib==3.9.4
multitasking==0.0.12
numpy==2.0.2
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pandas-datareader==0.10.0
pandas-stubs==2.0.3.230814
peewee==4.0.3
pillow==10.4.0
platformdirs==4.4.0
protobuf==6.33.6
psycopg2-binary==2.9.12
pycparser==2.23
PyJWT==2.0.0
pyparsing==3.3.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.4
selenium==4.12.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.5.3
tomli==2.4.0
trio==0.27.0
trio-websocket==0.12.2
types-pytz==2025.2.0.20251108
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20260107
typing_extensions==4.13.2
tzdata==2025.3
urllib3==2.6.3
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
yfinance==1.2.0
zipp==3.23.0
anthropic==0.98.1
